\name{bottom_up_spec}
\alias{bottom_up_spec}
\title{Create a bottom-UpR specification}
\usage{bottom_up_spec(data, response, buildings = NULL, response_type = c("auto", "count", "ppb"), likelihood = "auto", covariates = NULL, coords = c("x", "y"), random_effects = NULL, spatial = TRUE)}
\description{See the package README and vignette for details and examples.}
