\name{plot_hierarchical_effects}
\alias{plot_hierarchical_effects}
\title{Posterior hierarchical-effect plot}
\usage{plot_hierarchical_effects(object, effect = NULL, main = NULL)}
\description{Plots posterior means and 95 percent credible intervals for a selected hierarchical effect using original grouping labels.}
