\name{plot_spatial_field}
\alias{plot_spatial_field}
\title{Plot posterior spatial field}
\usage{plot_spatial_field(object, main = "Posterior spatial field")}
\arguments{
\item{object}{A spatial \code{bottom_up_fit}.}
\item{main}{Plot title.}
}
\value{Mesh-node coordinates and posterior means, invisibly.}
\description{Plots posterior spatial-field means at SPDE mesh nodes.}
