\name{plot_bottom_up_exploration}
\alias{plot_bottom_up_exploration}
\title{Exploratory population-data plots}
\usage{plot_bottom_up_exploration(data, response, buildings = NULL, covariates = NULL)}
\description{Produces compact exploratory plots for response, buildings and people-per-building structure.}
